requirements.txt
streamlit==1.28.1
numpy>=1.26.0,<2.2.0
tqdm==4.66.1
faiss-cpu==1.7.4
librosa==0.10.1
python-dotenv==1.0.0
soundfile==0.12.1
tensorflow==2.19.0
tensorflow-hub
